Move classes into a namespace alongside the controller
