+++ api/api-base/src/main/java/com/thoughtworks/go/api/serializers/CaseInsensitiveStringDeserializer.java
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.representers.config;
+package com.thoughtworks.go.api.serializers;
 
 import com.thoughtworks.go.config.CaseInsensitiveString;
 
 import java.util.function.Function;
 
-public class CaseInsensitiveStringSerializer implements Function<CaseInsensitiveString, String> {
+public class CaseInsensitiveStringDeserializer implements Function<String, CaseInsensitiveString> {
 
     @Override
-    public String apply(CaseInsensitiveString caseInsensitiveString) {
-        return caseInsensitiveString != null ? caseInsensitiveString.toString() : null;
+    public CaseInsensitiveString apply(String s) {
+        return new CaseInsensitiveString(s);
     }
 }
