+++ api/api-roles-config-v1/src/main/java/com/thoughtworks/go/apiv1/admin/security/RolesControllerV1Delegate.java
@@ -30,8 +30,8 @@ import com.thoughtworks.go.server.service.EntityHashingService;
 import com.thoughtworks.go.server.service.RoleConfigService;
 import com.thoughtworks.go.server.service.result.HttpLocalizedOperationResult;
 import com.thoughtworks.go.server.util.UserHelper;
-import gen.com.thoughtworks.go.representers.config.rolev1.RoleMapper;
-import gen.com.thoughtworks.go.representers.config.rolev1.RolesMapper;
+import gen.com.thoughtworks.go.apiv1.admin.security.representers.RoleMapper;
+import gen.com.thoughtworks.go.apiv1.admin.security.representers.RolesMapper;
 import org.springframework.http.HttpStatus;
 import spark.Request;
 import spark.Response;
